Make parsing more robust in the face of unknown or unexpected tags
Unfortunately it is common for certain tags to be present in the wrong
header, so there are many packages in the wild which rpm-rs would fail
to parse. Additionally any packages built with brand-new versions of
rpmbuild which contain tags not listed in our list of constants would
fail.

In either case, we ought to be a bit more tolerant. Users should be able
to parse the RPMs and do basic things with them even if the new
functionality isn't available.

closes #119