Remove algorithm dependence in signing/verifying interfaces
RPM only supports PGP signing. PGP signatures have a baked in algorithm.
So there's no need to bake the algorithm into the API and duplicate a
bunch of code when we can just verify at runtime that the parsed PGP
signature is one of the appropriate ones, and then do the appropriate
thing with it.