Change PGP signing to match rpm
When RPM signs packages it does not mark the PGP packets as 'critical'.
When packages signed with the current rpm crate are verified by older
RPM versions the Key ID can not be found and the verification fails.

This does not affect later versions of RPM since the PGP part has been
replaced with the sequoia library, which does not seem to be affected.

By changing the PGP pacets to match older RPM versions the rpm files are
now able to be verified by older versions.