Add content length to signature header
This adds the RPMTAG_SIGSIZE/RPMTAG_LONGSIGSIZE to the signature header,
which is required by some tooling to compute metadata about RPMs. Note
that the actual constants used are from the signature header constants:

https://github.com/rpm-software-management/rpm/blob/eb4998108638df7588ce9cb7ee26de2f97da691d/include/rpm/rpmtag.h#L443

The non-signature constants cannot be put in the signature headers
(the rpm tooling in the repo above throws an error when parsing such a
file). So it seems likely this is the right way to encode the content
length. It also avoids needing to update a header's value in place after
writing it (so that the content length would account for the header
indicating the length).

Documentation on these options:

* https://github.com/rpm-software-management/rpm/blob/master/docs/manual/tags.md#signatures-and-digests
* https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/8/html/packaging_and_distributing_software/new-features-in-rhel-8_packaging-and-distributing-software#ref_64-bit-rpm-tags_assembly_support-for-files-above-4-gb

The latter makes it explicit that this is the header + *compressed* payload.